Add convertor for form data to submit
