+++ src/components/loginForm/LoginForm.tsx
@@ -1,6 +1,7 @@
 import { zodResolver } from '@hookform/resolvers/zod';
 import { setTokenToLocalStorage } from '@middleware/authHandler';
 import { submitFormData } from '@middleware/formHandler';
+import { convertToLoginData } from '@middleware/helpers/dataConverter';
 import React from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
@@ -21,10 +22,7 @@ export default function LoginForm() {
 	});
 
 	const login = async (formData: LoginData) => {
-		const userLoginData = {
-			email: formData.email,
-			password: formData.password,
-		};
+		const userLoginData = convertToLoginData(formData);
 
 		const {
 			success,
