+++ src/components/registrationForm/RegistrationForm.tsx
@@ -3,6 +3,10 @@ import '@layouts/FormLayout.css';
 import { loginUser } from '@middleware/apiHandler';
 import { setTokenToLocalStorage } from '@middleware/authHandler';
 import { submitFormData } from '@middleware/formHandler';
+import {
+	convertToLoginData,
+	convertToRegistrationData,
+} from '@middleware/helpers/dataConverter';
 import React from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
@@ -23,12 +27,7 @@ export default function RegistrationForm() {
 	});
 
 	const submitForm = async (formData: RegistrationData) => {
-		const userToRegister = {
-			username: `${formData.firstName} ${formData.lastName}`,
-			email: formData.email,
-			password: formData.password,
-			role: formData.userType,
-		};
+		const userToRegister = convertToRegistrationData(formData);
 
 		const { success, error } = await submitFormData(
 			userToRegister,
@@ -38,10 +37,7 @@ export default function RegistrationForm() {
 
 		if (success) {
 			//ZALOGUJ OD RAZU
-			const userToLogin = {
-				email: formData.email,
-				password: formData.password,
-			};
+			const userToLogin = convertToLoginData(formData);
 			const loginResult = await loginUser(userToLogin);
 
 			if (loginResult.ok) {
